fix/UPDATE-advanced-ziplines.json
Calculator project example update:

Previous calculator had several bugs that were addressed. Per Quincy Larson's request, this file has been updated to provide a new calculator example existing under Free Code Camp's CodePen profile that has corrected the following bugs in the previous calculator example:

Entry box is now limited so all numbers / operations in use are visible.
Corrected the use of operators - now you can not begin with an operator or use an operator twice in a row.
Corrected the use of the decimal - now you can only use the decimal once per number to be evaluated.
Corrected the use of zero - you now may not begin with or use multiple zeros unless it is in relation to a number with a decimal.
Corrected a bug where once you received an answer, you may continue using the answer in another equation.
Corrected a Clear Entry (CE) bug where once an equation's answer is returned, it does not All Clear (AC) if 'CE' is clicked.