fix(challenges): update regular expression that fails (#56)
* fix(challenges): update regular expression that fails

Changes regular expression in testString to remove negative look behind that is not implemented
across all browsers.

ISSUES CLOSED: #55

* Fixed line spacing

It is definitely something to do with a linter. I re-ran the `npm run commit` in my own environment and watched the line spacing jump. Should probably be checked if it ends up happening to others. However, this should fix it all for this PR!

* Back to `prettier` formatting

Reverting back to pretty formatting from commit.