CSS: Add ::first-letter pseudo-element (#21009)
This is a basic introduction to the ::first-letter pseudo-element CSS selector. This document is meant to be a starting point and should be added to/expounded on.

P.S. I tried submitting a doc for the ::before selector, but apparently, it was already created, however, it's nowhere to be found in this directory (css/selectors/pseudo/). I wonder if a consolidation is in order 😛.