Work on the space-evenly option for flexbox justify-content (#37797)
This challenge requests setting the justify-content attribute to one of the possible values shown in the previous challenge. One of these values - space-evenly - doesn't actually work (at least for the English version).

In this commit I fix the space-evenly issues to some extent.

I cannot tick the box saying that all files changed are in the same world language. However, the changes made should be understandable in all languages.

There still is more to be done for languages other than English, like adding to the test text for Arabic, Portuguese, and Spanish, and adding descriptions for space-evenly in the previous challenge.