fix(learn): address escaped backticks (#40717)
* fix(learn): address escaped backticks

Addresses the instances of escaped backticks - where a backtick is
preceded by a backslash.  In most cases, this was left over from the
old parser. In some cases, a backtick was intended to be wrapped in
code tags and has been adjusted accordingly.

This issue came to light due to a bug in the translation flow on
Crowdin.

Signed-off-by: nhcarrigan <nhcarrigan@gmail.com>

* fix: EVEN MORE :( :( :(

Signed-off-by: nhcarrigan <nhcarrigan@gmail.com>

* fix: backslash nightmares

Signed-off-by: nhcarrigan <nhcarrigan@gmail.com>

* fix: When you wish upon a *******

Signed-off-by: nhcarrigan <nhcarrigan@gmail.com>

* fix(curriculum): md error introduced by formatter

* fix(curriculum): remove extra `s

* fix: restore quote symbol

Signed-off-by: nhcarrigan <nhcarrigan@gmail.com>

* fix: Typo

Co-authored-by: Oliver Eyton-Williams <ojeytonwilliams@gmail.com>

* fix: apply review changes

Applying review feedback from call with @RandellDawson.

Signed-off-by: nhcarrigan <nhcarrigan@gmail.com>

* fix: markdown does weird stuff sometimes

Can't stick backticks together - use code.

Signed-off-by: nhcarrigan <nhcarrigan@gmail.com>

Co-authored-by: Oliver Eyton-Williams <ojeytonwilliams@gmail.com>